a part of the web that can create a question with answers
The editor includes a minimum and maximum of answers [they can be modified with two variables] and it allows the user to choose which answer is the correct one. Furthermore,  the button to publish the question will only work when all the fields are filled and an answer is marked as correct.